cobra por hora tipo Double

+++ src/main/kotlin/ar/edu/unahur/obj2/tareas/Tarea.kt
@@ -57,6 +57,6 @@ class TareaIntegracion( responsable :  Trabajador): Tarea( responsable){
     fun agregarTarea(tareaAAgregar: Tarea) = tareasDentro.add(tareaAAgregar)
 }
 
-class  Trabajador (var cuantoCobraPorHora :  Int)
+class  Trabajador (var cuantoCobraPorHora :  Double)
